LPS-107004 Update integration tests
